Factor out BackgroundImagePicker component
