+++ packages/koenig-lexical/src/components/ui/cards/HeaderCard.jsx
@@ -1,11 +1,9 @@
 import KoenigNestedEditor from '../../KoenigNestedEditor';
 import PropTypes from 'prop-types';
 import React from 'react';
+import {BackgroundImagePicker} from '../BackgroundImagePicker';
 import {Button} from '../Button';
 import {ButtonGroupSetting, ColorPickerSetting, InputSetting, InputUrlSetting, SettingsDivider, SettingsPanel, ToggleSetting} from '../SettingsPanel';
-import {ReactComponent as FileUploadIcon} from '../../../assets/icons/kg-upload-fill.svg';
-import {ProgressBar} from '../ProgressBar';
-import {ReactComponent as TrashIcon} from '../../../assets/icons/kg-trash.svg';
 import {isEditorEmpty} from '../../../utils/isEditorEmpty';
 
 export const HEADER_COLORS = {
@@ -22,81 +20,6 @@ export const HEADER_TEXT_COLORS = {
     accent: 'text-white caret-white kg-header-accent',
     image: 'text-white caret-white'
 };
-
-function FileUploading({progress}) {
-    const progressStyle = {
-        width: `${progress?.toFixed(0)}%`
-    };
-
-    return (
-        <div className="h-full border border-transparent">
-            <div className="relative flex h-[120px] items-center justify-center border border-grey/20 bg-grey-50 before:pb-[12.5%] dark:bg-grey-900">
-                <div className="flex w-full items-center justify-center overflow-hidden">
-                    <ProgressBar style={progressStyle} />
-                </div>
-            </div>
-        </div>
-    );
-}
-
-function ImagePicker({onFileChange,
-    backgroundImageSrc,
-    handleClearBackgroundImage,
-    fileInputRef,
-    backgroundImagePreview,
-    openFilePicker,
-    isUploading,
-    progress}) {
-    if (isUploading) {
-        return (
-            <FileUploading progress={progress} />
-        );
-    }
-    return (
-        <>
-            <form onChange={onFileChange}>
-                <input
-                    ref={fileInputRef}
-                    accept='image/*'
-                    hidden={true}
-                    name="image-input"
-                    type='file'
-                />
-            </form>
-            {
-                backgroundImagePreview && (
-                    <div className="w-full">
-                        <div className="relative">
-                            <div className="flex w-full items-center justify-center">
-                                {
-                                    backgroundImageSrc ?
-                                        <>
-                                            <div className="group relative mb-4 w-full rounded">
-                                                <div className="absolute inset-0 rounded bg-gradient-to-t from-black/0 via-black/5 to-black/30 opacity-0 transition-all group-hover:opacity-100">
-                                                </div>
-                                                <div className="absolute top-5 right-5 flex opacity-0 transition-all group-hover:opacity-100">
-                                                    <button className="pointer-events-auto flex h-8 w-9 cursor-pointer items-center justify-center rounded bg-white/90 transition-all hover:bg-white" type="button" onClick={handleClearBackgroundImage}>
-                                                        <TrashIcon className="h-5 w-5 fill-grey-900 stroke-[3px] transition-all ease-linear group-hover:scale-105" />
-                                                    </button>
-                                                </div>
-                                                <img alt='backgroundHeaderImage' className="max-h-64 w-full rounded object-cover" data-testid="image-picker-background" src={backgroundImageSrc} />
-                                            </div>
-                                        </>
-                                        :
-                                        <button className="group flex h-[120px] w-full cursor-pointer flex-col items-center justify-center rounded-sm border border-dashed border-grey-300 bg-grey-50 dark:border-grey-800 dark:bg-grey-900" type="button" onClick={openFilePicker}>
-                                            <FileUploadIcon className="h-5 w-5 fill-grey-700 stroke-[3px] transition-all ease-linear group-hover:scale-105" />
-                                            <span className="px-1 text-[1.35rem] font-medium text-grey-700">Click to upload background image</span>
-                                        </button>
-                                }
-                            </div>
-                        </div>
-                    </div>
-                )
-            }
-
-        </>
-    );
-}
 
 export function HeaderCard({isEditing,
     size,
@@ -210,7 +133,7 @@ export function HeaderCard({isEditing,
                 }
 
                 {/* Button */}
-                { button ? 
+                { button ?
                     <div className={`${(size === 'small') ? 'mt-6' : (size === 'medium') ? 'mt-8' : 'mt-10'}`}>
                         {((button && (type === 'light')) && <Button dataTestId="header-card-button" placeholder={buttonPlaceholder} size={size} value={buttonText} />) || (button && <Button color='light' dataTestId="header-card-button" placeholder={buttonPlaceholder} size={size} value={buttonText} />)}
                     </div>
@@ -235,7 +158,7 @@ export function HeaderCard({isEditing,
                         selectedName={type}
                         onClick={handleColorSelector}
                     />
-                    <ImagePicker
+                    <BackgroundImagePicker
                         backgroundImagePreview={backgroundImagePreview}
                         backgroundImageSrc={backgroundImageSrc}
                         fileInputRef={fileInputRef}
@@ -307,19 +230,4 @@ HeaderCard.propTypes = {
     headerTextEditorInitialState: PropTypes.string,
     subheaderTextEditor: PropTypes.object,
     subheaderTextEditorInitialState: PropTypes.string
-};
-
-FileUploading.propTypes = {
-    progress: PropTypes.number
-};
-
-ImagePicker.propTypes = {
-    backgroundImagePreview: PropTypes.bool,
-    backgroundImageSrc: PropTypes.string,
-    fileInputRef: PropTypes.object,
-    handleClearBackgroundImage: PropTypes.func,
-    isUploading: PropTypes.bool,
-    openFilePicker: PropTypes.func,
-    progress: PropTypes.number,
-    onFileChange: PropTypes.func
 };
